add exclude-substrings flag (#16)
Previous version of performance CLI was working with file extensions, however now it should be run inside the Jenkins environment, at the same time Jenkinsfile will be generated from SDK that does not have such metadata as extension. Thus extensions flag was changed to exclude-substrings. Default values are set with the respect to a fixtures suite of every driver, excluding files with AST contents.

Signed-off-by: lwsanty <lwsanty@gmail.com>